+++ branches/sayonara-832/src/HelperStructs/CustomMimeData.h
@@ -51,7 +51,7 @@ public:
         _has_meta_data = (v_md.size() > 0);
 	}
 
-    uint getMetaData(MetaDataList& v_md) const {
+	size_t getMetaData(MetaDataList& v_md) const {
 
         if(!_has_meta_data) return 0;
         if(_v_md.size() == 0) return 0;
